raket-2/requirements.txt
requests==2.31.0
aiohttp==3.9.1
web3==6.11.1
pandas>=2.2.0
python-dotenv==1.0.0
aiolimiter==1.1.0
colorlog==6.7.0
asyncio==3.4.3
beautifulsoup4==4.12.2
eth-utils==2.2.0
tqdm>=4.65.0 